This PR makes sure that we can commit after a class changed shape, see #testMaterializingObjectWithIvarRemovedThenCommit
What was missing? We get the behavior description via #behaviorDescriptionFor: from the behaviorRegistry, but that one is the old one (with the ivar not removed)
Solution: create a new SOBehaviorDescription with the same objectID. 

This has to be moved to the registry later so we can find the history of a desciption, but it makes committing objects with changed classes possible now
The code in behaviorDescriptionFor: will be refactored later.

Small changes not related:

- testSOBehaviorDescription
- use SOBehaviorDescription meta", add #metaId that returns the id with index 2