The# isCurrent check for SoilBehaviorDescription is very slow: it creates a new intstance for the class in the system for each call.
This is not needed:

SoilBehaviorDescription instances are created per transaction (or for the materializer when standalone). 

As we can assume that class shape does not change during a tranaction (or the invokation of a materialize step).

We can therefore cache the result of the check in the instances, which just are alive for that transaction.